requirements.txt
# Qwen3-VL-4B Fine-Tuning MVP Dependencies
# Tested with Python 3.10+ / CUDA 12.1+

# Core ML
torch>=2.4.0
torchvision>=0.19.0

# Transformers (build from source for Qwen3-VL support)
transformers @ git+https://github.com/huggingface/transformers.git

# Fine-tuning stack
trl>=0.17.0
peft>=0.15.0
accelerate>=1.3.0

# Quantization (QLoRA)
bitsandbytes>=0.45.0

# Dataset
datasets>=3.0.0

# Qwen-VL image processing utilities
qwen-vl-utils>=0.0.8

# Misc
pillow>=10.0.0
tensorboard>=2.15.0

# Hardware monitoring
psutil>=5.9.0
pynvml>=11.5.0
matplotlib>=3.7.0